Small platformer_stock_B fix
- Resolved an error in the gravity function where negative values were not being added, as the elif statement would be the same as the if statement. 

CLARIFICATION ON NEGATIVE AND POSITIVE GRAVITY VALUES:
In this case, a negative value is the difference between the player and the block below. The negative gravity value would overwrite the positive gravity value if the positive value would put the player past the block boundary. Instead, the player will end up exactly on top of the block with the negative gravity value.